error_handling

+++ app/controllers/items_controller.rb
@@ -31,7 +31,11 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+    if item.seller_id == current_user.id
+      item.destroy
+    else
+      redirect_to item_path(params[:id])
+    end
   end
 
   def show
